added large output data storing

+++ ServerApp/Services/EndpointService.cs
@@ -9,6 +9,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
+using System.Text;
 using System.Text.RegularExpressions;
 
 namespace IntegrationTestingTool.Services
@@ -18,11 +19,14 @@ namespace IntegrationTestingTool.Services
         private IMongoCollection<Endpoint> MongoCollection { get; }
         private IConfigService ConfigService { get; }
 
-        public EndpointService(IDatabaseSettings settings, IConfigService configService)
+        private IFileService FileService { get; }
+
+        public EndpointService(IDatabaseSettings settings, IConfigService configService, IFileService fileService)
         {
             var client = new MongoClient(settings.ConnectionString);
             MongoCollection = client.GetDatabase(settings.DatabaseName).GetCollection<Endpoint>("Endpoints");
             ConfigService = configService;
+            FileService = fileService;
         }
 
         public IEnumerable<Endpoint> GetAll() =>
@@ -42,15 +46,26 @@ namespace IntegrationTestingTool.Services
         {
             endpoint.Id = Guid.NewGuid();
             endpoint.OutputData = Regex.Replace(endpoint.OutputData, @"\""", @"""");
-            MongoCollection.InsertOne(endpoint);
-            return endpoint;
-        }
 
+            var preprocessedEndpoint = HandleLargeOutputData(endpoint);
+            MongoCollection.InsertOne(preprocessedEndpoint);
+            return preprocessedEndpoint;
+        }
         public bool Delete(Guid id)
         {
-            var deletionFilter = Builders<Endpoint>.Filter.Eq(nameof(Endpoint.Id), id);
-            var result = MongoCollection.DeleteOne(deletionFilter);
-            return result.DeletedCount != 0;
+            var endpoint = FindById(id);
+            if (endpoint != null)
+            {
+                var deletionFilter = Builders<Endpoint>.Filter.Eq(nameof(Endpoint.Id), id);
+                var result = MongoCollection.DeleteOne(deletionFilter);
+                var isDeleted = result.DeletedCount != 0;
+                if (isDeleted)
+                {
+                    FileService.Delete(endpoint.OutputDataFile);
+                    return isDeleted;
+                }
+            }
+            return false;
         }
 
         public IEnumerable<Endpoint> FindByPathAndMethod(string path, string method)
@@ -59,6 +74,12 @@ namespace IntegrationTestingTool.Services
             var methodFilter = Builders<Endpoint>.Filter.Eq(nameof(Endpoint.Method), method);
             var filters = Builders<Endpoint>.Filter.And(pathFilter, methodFilter);
             return MongoCollection.Find(filters).ToList();
+        }
+
+        public Endpoint FindById(Guid id)
+        {
+            BsonBinaryData binaryId = new BsonBinaryData(id, GuidRepresentation.Standard);
+            return MongoCollection.Find(new BsonDocument("_id", binaryId)).FirstOrDefault();
         }
 
         public IEnumerable<Endpoint> FindByParameter(string parameterName, string value) =>
@@ -98,6 +119,27 @@ namespace IntegrationTestingTool.Services
         {
             var props = typeof(HttpMethod).GetProperties();
             return props.Where(prop => prop.GetMethod.IsStatic).Select(x => x.Name.ToUpper());
+        }
+
+        private Endpoint HandleLargeOutputData(Endpoint endpoint)
+        {
+            if (string.IsNullOrEmpty(endpoint.OutputData))
+            {
+                return endpoint;
+            }
+
+            endpoint.OutputDataSize = Encoding.UTF8.GetBytes(endpoint.OutputData).Length;
+
+            double fileSize = 10 * Math.Pow(2, 20);
+
+            //Store data into file it has size more than 10MB
+            if (endpoint.OutputDataSize > fileSize)
+            {
+                endpoint.OutputDataFile = FileService.Create(endpoint.Id, endpoint.OutputData);
+                endpoint.OutputData = null;
+            }
+
+            return endpoint;
         }
     }
 }
